+++ docs/javadoc/index.html
@@ -1,0 +1,79 @@
+<!DOCTYPE HTML>
+<html lang>
+<head>
+<!-- Generated by javadoc (22) on Sun Mar 23 19:06:38 MSK 2025 -->
+<title>Overview (CdrService 0.0.1-SNAPSHOT API)</title>
+<meta name="viewport" content="width=device-width, initial-scale=1">
+<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+<meta name="dc.created" content="2025-03-23">
+<meta name="description" content="package index">
+<meta name="generator" content="javadoc/PackageIndexWriter">
+<link rel="stylesheet" type="text/css" href="resource-files/jquery-ui.min.css" title="Style">
+<link rel="stylesheet" type="text/css" href="resource-files/stylesheet.css" title="Style">
+<script type="text/javascript" src="script-files/script.js"></script>
+<script type="text/javascript" src="script-files/jquery-3.6.1.min.js"></script>
+<script type="text/javascript" src="script-files/jquery-ui.min.js"></script>
+</head>
+<body class="package-index-page">
+<script type="text/javascript">var pathtoroot = "./";
+loadScripts(document, 'script');</script>
+<noscript>
+<div>JavaScript is disabled on your browser.</div>
+</noscript>
+<header role="banner">
+<nav role="navigation">
+<!-- ========= START OF TOP NAVBAR ======= -->
+<div class="top-nav" id="navbar-top"><button id="navbar-toggle-button" aria-controls="navbar-top" aria-expanded="false" aria-label="Toggle navigation links"><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span></button>
+<div class="skip-nav"><a href="#skip-navbar-top" title="Skip navigation links">Skip navigation links</a></div>
+<ul id="navbar-top-firstrow" class="nav-list" title="Navigation">
+<li class="nav-bar-cell1-rev">Overview</li>
+<li>Package</li>
+<li>Class</li>
+<li>Use</li>
+<li><a href="overview-tree.html">Tree</a></li>
+<li><a href="index-all.html">Index</a></li>
+<li><a href="help-doc.html#overview">Help</a></li>
+</ul>
+</div>
+<div class="sub-nav">
+<div id="navbar-sub-list"></div>
+<div class="nav-list-search"><a href="search.html">SEARCH</a>
+<input type="text" id="search-input" disabled placeholder="Search">
+<input type="reset" id="reset-button" disabled value="reset">
+</div>
+</div>
+<!-- ========= END OF TOP NAVBAR ========= -->
+<span class="skip-nav" id="skip-navbar-top"></span></nav>
+</header>
+<main role="main">
+<div class="header">
+<h1 class="title">CdrService 0.0.1-SNAPSHOT API</h1>
+</div>
+<div id="all-packages-table">
+<div class="caption"><span>Packages</span></div>
+<div class="summary-table two-column-summary">
+<div class="table-header col-first">Package</div>
+<div class="table-header col-last">Description</div>
+<div class="col-first even-row-color all-packages-table all-packages-table-tab1"><a href="ru/projects/CdrService/package-summary.html">ru.projects.CdrService</a></div>
+<div class="col-last even-row-color all-packages-table all-packages-table-tab1">&nbsp;</div>
+<div class="col-first odd-row-color all-packages-table all-packages-table-tab1"><a href="ru/projects/CdrService/controller/package-summary.html">ru.projects.CdrService.controller</a></div>
+<div class="col-last odd-row-color all-packages-table all-packages-table-tab1">&nbsp;</div>
+<div class="col-first even-row-color all-packages-table all-packages-table-tab1"><a href="ru/projects/CdrService/model/package-summary.html">ru.projects.CdrService.model</a></div>
+<div class="col-last even-row-color all-packages-table all-packages-table-tab1">&nbsp;</div>
+<div class="col-first odd-row-color all-packages-table all-packages-table-tab1"><a href="ru/projects/CdrService/repository/package-summary.html">ru.projects.CdrService.repository</a></div>
+<div class="col-last odd-row-color all-packages-table all-packages-table-tab1">&nbsp;</div>
+<div class="col-first even-row-color all-packages-table all-packages-table-tab1"><a href="ru/projects/CdrService/service/package-summary.html">ru.projects.CdrService.service</a></div>
+<div class="col-last even-row-color all-packages-table all-packages-table-tab1">&nbsp;</div>
+<div class="col-first odd-row-color all-packages-table all-packages-table-tab1"><a href="ru/projects/CdrService/util/package-summary.html">ru.projects.CdrService.util</a></div>
+<div class="col-last odd-row-color all-packages-table all-packages-table-tab1">&nbsp;</div>
+<div class="col-first even-row-color all-packages-table all-packages-table-tab1"><a href="ru/projects/CdrService/util/serialization/package-summary.html">ru.projects.CdrService.util.serialization</a></div>
+<div class="col-last even-row-color all-packages-table all-packages-table-tab1">&nbsp;</div>
+</div>
+</div>
+</main>
+<footer role="contentinfo">
+<hr>
+<p class="legal-copy"><small>Copyright &#169; 2025. All rights reserved.</small></p>
+</footer>
+</body>
+</html>
