+++ src/main/java/ru/projects/CdrService/model/Subscriber.java
@@ -5,6 +5,9 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 
+/**
+ * Класс модели абонента
+ */
 @Entity
 public class Subscriber {
     @Id
