+++ src/main/java/ru/projects/CdrService/service/UdrService.java
@@ -12,6 +12,11 @@ import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * Сервис для работы с UDR-записями абонентов
+ * @see CdrRepository
+ * @see SubscriberRepository
+ */
 @Service
 public class UdrService {
     @Autowired
@@ -20,6 +25,15 @@ public class UdrService {
     @Autowired
     private SubscriberRepository subscriberRepository;
 
+    /**
+     * Возвращает UDR-запись за запрошенный месяц или за весь год
+     * @param msisdn Номер абонента
+     * @param month Месяц, за который необходимо получить UDR-запись(1-12). Если указано число меньше 1 или больше 12,
+     *              то UDR-запись составляется за весь год
+     * @return Возвращает UDR-запись за запрошенный месяц, или за весь год, или null, если абонента с номером {@param msisdn}
+     *      не существует
+     * @see Udr
+     */
     public Udr getUdrByOneSubscriber(String msisdn, int month) {
         if (!subscriberRepository.existsByMsisdn(msisdn)) {
             return null;
@@ -34,7 +48,7 @@ public class UdrService {
 
         udr.setMsisdn(msisdn);
         udr.setIncomingCall(new Udr.CallDuration());
-        udr.setOutcomingCall(new Udr.CallDuration());
+        udr.setOutgoingCall(new Udr.CallDuration());
         for (var cdr: cdrRecords) {
 
             Duration duration = Duration.between(cdr.getStartCall(), cdr.getEndCall());
@@ -42,18 +56,20 @@ public class UdrService {
             if (cdr.getCallerNumber().equals(msisdn)) {
                 udr.getIncomingCall().setTotalTime(udr.getIncomingCall().getTotalTime().plus(duration));
             } else {
-                udr.getOutcomingCall().setTotalTime(udr.getOutcomingCall().getTotalTime().plus(duration));
+                udr.getOutgoingCall().setTotalTime(udr.getOutgoingCall().getTotalTime().plus(duration));
             }
         }
 
         return udr;
     }
 
+    /**
+     * Возвращает список UDR записей для всех абонентов за указанный месяц
+     * @param month Месяц, за который необходимо получить UDR-записи(1-12).
+     * @return Список UDR-записей за запрошенный месяц
+     * @see Udr
+     */
     public List<Udr> getUdrByAllSubscribers(int month) {
-        if (month < 1 || month > 12) {
-            return null;
-        }
-
         List<Udr> udrRecords = new ArrayList<>();
         List<String> msisdnList = subscriberRepository.findAll().stream()
                 .map(Subscriber::getMsisdn)
