+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
 spring.application.name=CdrService
-server.port=8000
+server.port=8088
 
 # H2 Database configuration
 spring.datasource.url=jdbc:h2:mem:cdr_service_db
