+++ src/test/java/ru/projects/CdrService/service/SwitchServiceTest.java
@@ -1,0 +1,22 @@
+package ru.projects.CdrService.service;
+
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+import ru.projects.CdrService.repository.CdrRepository;
+import ru.projects.CdrService.repository.SubscriberRepository;
+
+@ExtendWith(MockitoExtension.class)
+public class SwitchServiceTest {
+    @Mock
+    CdrRepository cdrRepository;
+
+    @Mock
+    SubscriberRepository subscriberRepository;
+
+    @InjectMocks
+    SwitchService switchService;
+
+
+}
